requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap4==2024.10
Django==5.2
django-crispy-forms==2.4
idna==3.10
pillow==11.2.1
pypng==0.20220715.0
pytz==2025.2
qrcode==8.1
requests==2.32.3
rrdtool-bindings==0.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
django-allauth>=0.63.0
cryptography>=42.0.0
PyJWT>=2.8.0
authlib>=1.2.0
